Move custom comparison to own file
